Silence a warning about comparing NSUInteger and unsigned int on 64-bit architectures.
git-svn-id: https://vrjuggler.googlecode.com/svn/juggler/trunk@21134 08b38cba-cd3b-11de-854e-f91c5b6e4272

+++ modules/vrjuggler/vrj/Kernel/VRJBasicDelegate.mm
@@ -234,7 +234,7 @@ NSString* VRJRecentCfgFiles = @"VRJRecentCfgFiles";
 
       // Update the submenu listing the recently opened configuration files
       // to include fileName.
-      const unsigned int index = [mRecentCfgFiles indexOfObject:fileName];
+      const NSUInteger index = [mRecentCfgFiles indexOfObject:fileName];
 
       if ( index != NSNotFound )
       {
